backend/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.16,<3.17
python-dotenv>=1.1,<2.0
psycopg[binary]>=3.2,<4.0
